Separated and fixed debug + verify test attempting to otherwise close a server instance that didn't start.
