+++ test/test-examples-server-app-loading-fail-checks.js
@@ -23,20 +23,6 @@ describe("Alexa App Server with Examples & App loading fail checking", function(
       .get('/')
       .expect(200).then(function(response) {
         expect(response.text).to.contain("alexa-app-server is running");
-      }
-    );
-  });
-
-  it("should throw an error when 'debug' and 'verify' are enabled", function() {
-    var fn = function() {
-      testServer = alexaAppServer.start({
-        port: 3000,
-        server_root: 'invalid_examples',
-        debug: true,
-        verify: true
       });
-    };
-
-    expect(fn).to.throw(Error);
   });
 });
